Fix type confusion (size_t)

+++ src/saveconv.c
@@ -274,13 +274,13 @@ static bool savetype_is_moo13(struct game_s *g, const char *fname)
     } while (0)
 #define M13_GET_TBL_16(item_, addr_) \
     do { \
-        for (int i_ = 0; i_ < TBLLEN(item_); ++i_) { \
+        for (size_t i_ = 0; i_ < TBLLEN(item_); ++i_) { \
             item_[i_] = GET_LE_16(&save2buf[(addr_) + i_ * 2]); \
         } \
     } while (0)
 #define M13_GET_TBL_16_OWNER(item_, addr_) \
     do { \
-        for (int i_ = 0; i_ < TBLLEN(item_); ++i_) { \
+        for (size_t i_ = 0; i_ < TBLLEN(item_); ++i_) { \
             uint16_t t_; \
             t_ = GET_LE_16(&save2buf[(addr_) + i_ * 2]); \
             if (t_ == 0xffff) { t_ = PLAYER_NONE; }; \
@@ -289,7 +289,7 @@ static bool savetype_is_moo13(struct game_s *g, const char *fname)
     } while (0)
 #define M13_GET_TBL_16_HATED(item_, addr_) \
     do { \
-        for (int i_ = 0; i_ < TBLLEN(item_); ++i_) { \
+        for (size_t i_ = 0; i_ < TBLLEN(item_); ++i_) { \
             uint16_t t_; \
             t_ = GET_LE_16(&save2buf[(addr_) + i_ * 2]); \
             if (t_ == 0) { t_ = PLAYER_NONE; } \
@@ -326,7 +326,7 @@ static bool savetype_is_moo13(struct game_s *g, const char *fname)
     } while (0)
 #define M13_GET_TBL_32(item_, addr_) \
     do { \
-        for (int i_ = 0; i_ < TBLLEN(item_); ++i_) { \
+        for (size_t i_ = 0; i_ < TBLLEN(item_); ++i_) { \
             item_[i_] = GET_LE_32(&save2buf[(addr_) + i_ * 4]); \
         } \
     } while (0)
@@ -740,7 +740,7 @@ static int savetype_de_moo13(struct game_s *g, const char *fname)
     } while (0)
 #define M13_SET_TBL_16(item_, addr_) \
     do { \
-        for (int i_ = 0; i_ < TBLLEN(item_); ++i_) { \
+        for (size_t i_ = 0; i_ < TBLLEN(item_); ++i_) { \
             uint16_t t_; \
             t_ = item_[i_]; \
             SET_LE_16(&save2buf[(addr_) + i_ * 2], t_); \
@@ -748,7 +748,7 @@ static int savetype_de_moo13(struct game_s *g, const char *fname)
     } while (0)
 #define M13_SET_TBL_16_OWNER(item_, addr_) \
     do { \
-        for (int i_ = 0; i_ < TBLLEN(item_); ++i_) { \
+        for (size_t i_ = 0; i_ < TBLLEN(item_); ++i_) { \
             uint16_t t_; \
             t_ = item_[i_]; \
             if (t_ == PLAYER_NONE) { t_ = 0xffff; }; \
@@ -757,7 +757,7 @@ static int savetype_de_moo13(struct game_s *g, const char *fname)
     } while (0)
 #define M13_SET_TBL_16_HATED(item_, addr_) \
     do { \
-        for (int i_ = 0; i_ < TBLLEN(item_); ++i_) { \
+        for (size_t i_ = 0; i_ < TBLLEN(item_); ++i_) { \
             uint16_t t_; \
             t_ = item_[i_]; \
             if (t_ == PLAYER_NONE) { t_ = 0; } \
@@ -790,7 +790,7 @@ static int savetype_de_moo13(struct game_s *g, const char *fname)
     } while (0)
 #define M13_SET_TBL_32(item_, addr_) \
     do { \
-        for (int i_ = 0; i_ < TBLLEN(item_); ++i_) { \
+        for (size_t i_ = 0; i_ < TBLLEN(item_); ++i_) { \
             uint32_t t_; \
             t_ = item_[i_]; \
             SET_LE_32(&save2buf[(addr_) + i_ * 4], t_); \
